# JText::_() and JText::loadLanguage() Optimisations