@andrepereiradasilva 2016-03-12
At least loadLanguage performance has improved in 3.5.0 RC so is now faster.
See https://github.com/joomla/joomla-cms/blob/staging/libraries/joomla/language/language.php#L798-L805