@brianteeman 2016-07-21
Hello @orware

Could you please let us know if this pull request is still valid?
If no reply is received within 4 weeks then we will close this issue.
<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/6788">issues.joomla.org/joomla-cms/6788</a>.</sub>