@orware 2016-07-21
Just took a quick revisit of the PR and the current version of the file:
https://github.com/joomla/joomla-cms/blob/staging/libraries/joomla/language/language.php

I see what izharaazmi mentions above about array_replace being used (which covers 2/3 of the lines changed in my original PR). If that's already being done as a performance improvement then that probably invalidates my original code there.

The only other part that might still be applicable is the if statement I added around line 371 described in my comment about "JText::_() Changes" above which will still save some computing effort if implemented.

Before Patch (5000 Calls, Time = 35.2 ms):
Time: 25.6 ms / 26.4 ms Memory: 2.136 MB / 2.83 MB Application: Before JText::_()
Time: 35.2 ms / 61.7 ms Memory: 0.288 MB / 3.12 MB Application: After JText::_()

After Patch (5000 Calls, Time = 28.2 ms):
Time: 27.0 ms / 27.9 ms Memory: 2.136 MB / 2.83 MB Application: Before JText::_()
Time: 28.2 ms / 56.1 ms Memory: 0.289 MB / 3.12 MB Application: After JText::_()<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/6788">issues.joomla.org/joomla-cms/6788</a>.</sub>