@orware 2016-07-21
Hi andrepereiradasilva,

Just created a new PR for this with just those changes:
https://github.com/joomla/joomla-cms/pull/11235

With a link back to this issue.<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/6788">issues.joomla.org/joomla-cms/6788</a>.</sub>