@brianteeman 2016-07-21
Closed as we have a new PR -Thanks<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/6788">issues.joomla.org/joomla-cms/6788</a>.</sub>