@gunjanpatel 2016-07-22
Set to "closed" on behalf of @gunjanpatel by The <a href="https://github.com/joomla/jissues">JTracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/6788">issues.joomla.org/joomla-cms/6788</a>